Match cases by "style" instead of datetime
To be able to notice when court dates move we should use the primary
key:

  (case number, hearing type, style)

One of these will be different if there are multiple events. For
instance, the hearing type will be either Arraignment or Hearing if
there are two scheduled cases.

The trick is if there are multiple days of a trial scheduled, I think
this will break because they are all the same. TODO for now!