Refactor diffing logic into its own module
This is a sizeable refactor of responsibility, by extracting the court
case diffing logic into a `Differ` module that takes two arrays:

1. `@before` which are the court cases from the database (they have an
   `:id` attribute)
2. `@after` which are the court cases from the latest scraping of the
   page

By extracting the responsibility for diffing these two arrays into a
module it allows us to write some tests of various conditions and assert
that the correct event is emitted (i.e. a `:changed` event instead of a
`:removed` and `:added` pair of events).