Add tests for ScheduleUpdater
This module is pretty hard to test due to its dependencies and logic.
But it is one of the most important!

I refactored it a little bit to move the new_schedule fetching logic
into its own method (#fetch_schedules) so that I can break the test
suite into two chunks (one of each method: fetch_schedules and
update) with each stubbing the other.

There's more to test, but this is a great starting point!