Fix datetime formatting in created message
Importantly, the format string was wrong, effectively rendering HH:HH
instead of HH:MM for the time part. But also I had forgotten to actually
apply the format string to the datetime itself, instead just sending out
the raw format string.

Because we have tests for this now, I was able to write a test for it
(and that is actually how I caught that the format string was wrong!) To
make the test suite independent of the processes's time zone, I made
sure that the schedule is created with a UTC datetime instead of a local
one.